refactor: move validation folder to utils
